made it isocortex specific
This corrects the issue with hippocampus being found by adding an extra filter to make sure returned layers are in isocortex. Still needs to be able to use the native regex and not the accessory one provided, though. There appears to be a function for get_layer_mask that would probably be more elegant to use.