Excitatory split (#19)
* Write SSCX layer-specific types to all isocortex

Changed layer key in csv file, changed layer filter in metadata to make it specific for Isocortex layers, changed code to do isocortex instead of just SSCX

* made it isocortex specific

This corrects the issue with hippocampus being found by adding an extra filter to make sure returned layers are in isocortex. Still needs to be able to use the native regex and not the accessory one provided, though. There appears to be a function for get_layer_mask that would probably be more elegant to use.

Used code snippet from Dimitri to read in metadata and create layers. It seems to filter okay for isocortex. I foresee the metadata.json file being put into the CLL, so for now it's included as a helper file but eventually would be specified.

Co-authored-by: dkeller9 <dkeller9@gmail.com>